Refactor into main and game class files
